requirements.txt
opencv-python-headless==4.5.1.48
numpy~=1.19.5
scipy~=1.5.4
scikit-image~=0.18.0rc1
matplotlib~=3.3.4
Flask~=1.1.2
wsq~=0.3
Pillow~=8.1.0
imutils~=0.5.4
CMake~=3.21.4
dlib~=19.22.1
face_recognition~=1.3.0
gunicorn==20.0.4
flask_cors~=3.0.10
dnspython~=2.1.0
pymongo~=3.11.3
requests~=2.25.1